requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
idna==3.6
jieba3k==0.35.1
joblib==1.3.2
lxml==4.9.3
newspaper3k==0.2.8
nltk==3.8.1
Pillow==10.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
textblob==0.17.1
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
urllib3==2.1.0
